% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VTShiny.R
\name{VTShiny}
\alias{VTShiny}
\title{Launch 'VTShiny' Interface}
\usage{
VTShiny()
}
\value{
Nothing
}
\description{
VTShiny() loads interactive user interface built using R 'shiny'.
}
\details{
Start VTShiny

The interactive user interface is to provide an easy way for variance analysis.
}
\examples{
if(interactive()){
library(rmarkdown)
VTShiny()
}
}
\keyword{VTShiny}
